VectorSpinVlasovMaxwell1D1V

Documentation for VectorSpinVlasovMaxwell1D1V.

VectorSpinVlasovMaxwell1D1V.BSplineAdvection

VectorSpinVlasovMaxwell1D1V.Mesh

VectorSpinVlasovMaxwell1D1V.H1f!

VectorSpinVlasovMaxwell1D1V.H2fh!

VectorSpinVlasovMaxwell1D1V.H3fh!

VectorSpinVlasovMaxwell1D1V.HAA!

VectorSpinVlasovMaxwell1D1V.He!

VectorSpinVlasovMaxwell1D1V.advection!

VectorSpinVlasovMaxwell1D1V.bspline

VectorSpinVlasovMaxwell1D1V.numeint

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.translation!

VectorSpinVlasovMaxwell1D1V.BSplineAdvection — Type

struct BSplineAdvection

Advection to be computed on each row

dims::Symbol

p::Int64

step::Float64

modes::Vector{Float64}

eig_bspl::Vector{Float64}

eigalpha::Vector{ComplexF64}

fhat::Matrix{ComplexF64}

VectorSpinVlasovMaxwell1D1V.Mesh — Type

struct Mesh

Mesh type to store domain parameters

nv::Int64

Number of points in v

nx::Int64

Number of points in x

vmin::Float64

Domain size v ∈ ]vmin,vmax[

vmax::Float64

Domain size v ∈ ]vmin,vmax[

xmin::Float64

Domain size x ∈ [xmin,xmax]

xmax::Float64

Domain size x ∈ [xmin,xmax]

kx::Vector{Float64}

Wave number vector to compute derivative with FFTs

dx::Float64

Size step along x

dv::Float64

Size step along v

x::Vector{Float64}

points along x direction

v::Vector{Float64}

points along v direction

vnode::Vector{Float64}

center points along v direction

VectorSpinVlasovMaxwell1D1V.H1f! — Method

H1f!(f0, f1, f2, f3, E1, t, L, H)

VectorSpinVlasovMaxwell1D1V.H2fh! — Method

compute the subsystem H2

H2fh!(f0, f1, f2, f3, E3, A3, t, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.H3fh! — Method

H3fh!(f0, f1, f2, f3, E2, A2, t, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.HAA! — Method

HAA!(f0, f1, f2, f3, E2, E3, A2, A3, t, L, H)

VectorSpinVlasovMaxwell1D1V.He! — Method

He!(f0, f1, f2, f3, E1, E2, E3, A2, A3, t, H)

compute the first subsystem He()

M is odd number

VectorSpinVlasovMaxwell1D1V.advection! — Method

advection!(df, adv, v, dt)

interpolate df(x - delta) with Parabolic Spline Method (PSM)

We consider a linear advection problem in $p$ direction

\[\frac{\partial f}{\partial t} + a \frac{\partial f}{\partial x} =0.\]

From the conservation of the volume, we have the following identity

\[f_{j,\ell}(t)=\frac{1}{\Delta p} \int_{p_{\ell-1/2}} ^{p_{\ell+1/2}} f(x_j,p,t)\mathrm{d}{p} =\frac{1}{\Delta p} \int_{p_{\ell-1/2}-at} ^{p_{\ell+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

For simplicity, denote by $q\in [1, M]$ the index such that $p_{\ell+1/2}-at \in [p_{q-1/2},p_{q+1/2}]$ i.e. $p_{\ell+1/2}-at \in C_q$, then we have

\[f_{j,\ell}(t) =\frac{1}{\Delta p} \int_{p_{q-1/2}-at} ^{p_{q-1/2}} f(x_j,p,0)\mathrm{d}{p}+f_{j,q}(0)-\frac{1}{\Delta p} \int_{p_{q+1/2}} ^{p_{q+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

Here we need to reconstruct a polynomial function $f(x_j,p,0)$ using the averages $f_{j,l}(0)$ using the PSM approach.

VectorSpinVlasovMaxwell1D1V.bspline — Method

bspline(p, j, x)

Return the value at x in [0,1] of the B-spline with integer nodes of degree p with support starting at j. Implemented recursively using the De Boor's Algorithm

\[B_{i,0}(x) := \left\{ \begin{matrix} 1 & \mathrm{if} \quad t_i ≤ x < t_{i+1} \\ 0 & \mathrm{otherwise} \end{matrix} \right.\]

\[B_{i,p}(x) := \frac{x - t_i}{t_{i+p} - t_i} B_{i,p-1}(x) + \frac{t_{i+p+1} - x}{t_{i+p+1} - t_{i+1}} B_{i+1,p-1}(x).\]

VectorSpinVlasovMaxwell1D1V.numeint — Method

computation of integral average in each cell using newton-cotes formula

numeint(value, N)

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E2, A2, op, dt, h_int)

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E1, E2, E3, A2, A3, op, dt)

compute the first subsystem He()

M is odd number

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E2, E3, A2, A3, op, dt)

VectorSpinVlasovMaxwell1D1V.step! — Method

compute the subsystem H2

step!(f0, f1, f2, f3, E3, A3, op, dt, h_int)

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E1, op, dt)

VectorSpinVlasovMaxwell1D1V.translation! — Method

translation!(df, delta, H)

interpolate df(x - delta)