VectorSpinVlasovMaxwell1D1V

Documentation for VectorSpinVlasovMaxwell1D1V.

VectorSpinVlasovMaxwell1D1V.Mesh

VectorSpinVlasovMaxwell1D1V.H1f

VectorSpinVlasovMaxwell1D1V.H1f!

VectorSpinVlasovMaxwell1D1V.H2fh

VectorSpinVlasovMaxwell1D1V.H2fh!

VectorSpinVlasovMaxwell1D1V.H3fh

VectorSpinVlasovMaxwell1D1V.H3fh!

VectorSpinVlasovMaxwell1D1V.HAA

VectorSpinVlasovMaxwell1D1V.HAA!

VectorSpinVlasovMaxwell1D1V.He

VectorSpinVlasovMaxwell1D1V.He!

VectorSpinVlasovMaxwell1D1V.numeint

VectorSpinVlasovMaxwell1D1V.recover

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.step!

VectorSpinVlasovMaxwell1D1V.translation

VectorSpinVlasovMaxwell1D1V.translation!

VectorSpinVlasovMaxwell1D1V.translation!

VectorSpinVlasovMaxwell1D1V.Mesh — Type

struct Mesh

Mesh type to store domain parameters

N::Int64

Number of points in v

M::Int64

Number of points in x

H::Float64

Domain size v ∈ ]-H,+H[

L::Float64

Domain size x ∈ [0,L]

k::Vector{Float64}

Wave number vector to compute derivative with FFTs

dx::Float64

Size step along x

dv::Float64

Size step along v

x::Vector{Float64}

points along x direction

v::Vector{Float64}

points along v direction

VectorSpinVlasovMaxwell1D1V.H1f! — Method

H1f!(f0, f1, f2, f3, E1, t, L, H)

VectorSpinVlasovMaxwell1D1V.H1f — Method

H1f(f0, f1, f2, f3, E1, t, M, N, L, H)

VectorSpinVlasovMaxwell1D1V.H2fh! — Method

compute the subsystem H2

H2fh!(f0, f1, f2, f3, E3, A3, t, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.H2fh — Method

compute the subsystem H2

H2fh(f0, f1, f2, f3, E3, A3, t, M, N, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.H3fh! — Method

H3fh!(f0, f1, f2, f3, E2, A2, t, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.H3fh — Method

H3fh(f0, f1, f2, f3, E2, A2, t, M, N, L, H, h_int)

VectorSpinVlasovMaxwell1D1V.HAA! — Method

HAA!(f0, f1, f2, f3, E2, E3, A2, A3, t, L, H)

VectorSpinVlasovMaxwell1D1V.HAA — Method

HAA(f0, f1, f2, f3, E2, E3, A2, A3, t, M, N, L, H)

VectorSpinVlasovMaxwell1D1V.He! — Method

He!(f0, f1, f2, f3, E1, E2, E3, A2, A3, t, H)

compute the first subsystem He()

M is odd number

VectorSpinVlasovMaxwell1D1V.He — Method

He(f0, f1, f2, f3, E1, E2, E3, A2, A3, t, M, N, H)

compute the first subsystem He()

M is odd number

VectorSpinVlasovMaxwell1D1V.numeint — Method

computation of integral average in each cell using newton-cotes formula

numeint(value, N)

VectorSpinVlasovMaxwell1D1V.recover — Method

recover(f, N)

Given integral average in each cell;this function could compute coefficients a;b;c of piecewise quadratic polynomial using PSM method

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E2, A2, op, t, h_int)

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E1, E2, E3, A2, A3, op, t)

compute the first subsystem He()

M is odd number

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E2, E3, A2, A3, op, t)

VectorSpinVlasovMaxwell1D1V.step! — Method

compute the subsystem H2

step!(f0, f1, f2, f3, E3, A3, op, t, h_int)

VectorSpinVlasovMaxwell1D1V.step! — Method

step!(f0, f1, f2, f3, E1, op, t)

VectorSpinVlasovMaxwell1D1V.translation! — Method

translation!(df, delta, H)

interpolate df(x - delta)

VectorSpinVlasovMaxwell1D1V.translation! — Method

translation!(df, adv, delta)

interpolate df(x - delta) with Parabolic Spline Method (PSM)

We consider a linear advection problem in $p$ direction

\[\frac{\partial f}{\partial t} + a \frac{\partial f}{\partial x} =0.\]

From the conservation of the volume, we have the following identity

\[f_{j,\ell}(t)=\frac{1}{\Delta p} \int_{p_{\ell-1/2}} ^{p_{\ell+1/2}} f(x_j,p,t)\mathrm{d}{p} = rac{1}{\Delta p} \int_{p_{\ell-1/2}-at} ^{p_{\ell+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

For simplicity, denote by $q\in [1, M]$ the index such that $p_{\ell+1/2}-at \in [p_{q-1/2},p_{q+1/2}]$ i.e. $p_{\ell+1/2}-at \in C_q$, then we have

\[f_{j,\ell}(t) =\frac{1}{\Delta p} \int_{p_{q-1/2}-at} ^{p_{q-1/2}} f(x_j,p,0)\mathrm{d}{p}+f_{j,q}(0)-\frac{1}{\Delta p} \int_{p_{q+1/2}} ^{p_{q+1/2}-at} f(x_j,p,0)\mathrm{d}{p}.\]

Here we need to reconstruct a polynomial function $f(x_j,p,0)$ using the averages $f_{j,l}(0)$ using the PSM approach.

VectorSpinVlasovMaxwell1D1V.translation — Method

translation(oldvector, N, delta, H)

oldvector is the integral average value in each cell of function f(x) newvector is the integral average value in each cell of function f(x+delta)